data_processor/pyproject.toml
[tool.poetry]
name = "data-processor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "data_processor"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.1"
google-api-python-client = "^2.39.0"
oauth2client = "^4.1.3"
python-dotenv = "^0.19.2"
requests = "^2.27.1"
google-auth = "^2.6.2"
python-dateutil = "^2.8.2"
google-analytics-data = "^0.11.2"
google-auth-oauthlib = "^0.5.1"
boto3 = "^1.23.8"
rq = "^1.11.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pytest = "^7.0.1"
pytest-mock = "^3.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
